Set queryselect value on tutorial start
