+++ context/app/static/js/components/cells/QuerySelect/QuerySelect.jsx
@@ -1,27 +1,14 @@
-import React, { useState } from 'react';
+import React from 'react';
 import MenuItem from '@material-ui/core/MenuItem';
 import Button from '@material-ui/core/Button';
 
 import { queryTypes } from 'js/components/cells/queryTypes';
 import { capitalizeString } from 'js/helpers/functions';
-import { useStore } from 'js/components/cells/store';
 import { StyledTextField } from './style';
-
-const cellsStoreSelector = (state) => state.setQueryType;
+import { useQuerySelect } from './hooks';
 
 function QuerySelect({ completeStep, setParametersButtonRef }) {
-  const [selectedQueryType, setSelectedQueryType] = useState(queryTypes.gene.value);
-
-  const setQueryType = useStore(cellsStoreSelector);
-
-  function handleSelect(event) {
-    setSelectedQueryType(event.target.value);
-  }
-
-  function handleButtonClick() {
-    completeStep(`${capitalizeString(selectedQueryType)} Query`);
-    setQueryType(selectedQueryType);
-  }
+  const { selectedQueryType, handleSelect, handleButtonClick } = useQuerySelect(completeStep);
   return (
     <div>
       <StyledTextField
